serde for Cert: switch to #[serde(with="mod")] method
instead of doing a bunch of weird stuff to get a newtype Ser/De
implementation and then convert awkwardly between the newtype and
Cert, use the #[serde(with="mod")] attribute and implement the methods
for Vec<Cert>, as that's the type of every field involving certs which
we currently serialize.

Had to likewise manually derive Hash for KeyGenRequest as a whole
since Cert doesn't implement Hash either.

TestData still uses Vec<String> and manual conversion for certs with
private key data as sequoia's cert serialization strips secret key
data unless you switch to a TSK.  We want to preserve this behavior in
the normal case, so just special-case TestData for now as it's the
only place we serialize secrets.

Code is now substantially simpler than when using a newtype to
implement Ser/De/Hash.  The scaffolding will be removed in a separate
commit.