KeyConfig/generate: add outside revocation key handling
This commit requires regeneration of test data:
    rm -f test_data/*
    cargo run --bin generate

Add the ability to specify certs which can revoke keys issued
by the generated cert.

Likely to be useful in sharksign workflows so that keys can be revoked
by trusted individuals (or a different set of individuals, if the
submitted revocation certs are themselves managed by sharksign) in a
timely manner, rather than requiring regeneration of the generated
revocation key.