remove duplication of test_data.rs
Ah. So that's how "use super::foo" works; it's not looking at the
filesystem hierarchy for 'foo.rs' like I assumed, it's looking at the
module that imported it with "mod bar", and checking the namespace for
'foo' (same as a submodule within a file does -- in retrospect it's
obvious).

use-ing super::pgp::Cert failed because neither generate.rs nor
http.rs brought 'pgp' into scope, but it worked for src/lib.rs (and
src/sharksign/mod.rs before it) because that file did bring 'pgp' into
scope.  The data module already re-exports Cert, so we can use that
path to the identifier in test_data to avoid imposing more
restrictions on the caller.

This is all more brittle than I'd like, but it seems to work fine.