refactor App creation into its own function
Since one of App's type parameters is a private type, we
can't return the App directly (at least not as an App
rather than as an impl trait). Both callers currently
expect an IntoServiceFactory<S>, and the simplest way
to get an impl IntoServiceFactory<S> was to convert
the App directly into a ServiceFactory within the
function itself and return an impl ServiceFactory
with all the necessary bounds set.

I don't want to talk about how long this took to figure
out.

Additionally: fix the service functions in main; it seems
like using multiple scopes does not work (or at least
doesn't work like I assumed it did).

As of this commit main continues to not work due to an
interdependency issue between actix-web and actix-rt; that
will be fixed by another commit.